+++ tables/index.py
@@ -307,13 +307,13 @@ class Index(NotLoggedMixin, Group, indexesextension.Index):
             # bucket outside of a slice.
             maxnb = 2**8
             if self.slicesize > maxnb * lbucket:
-                lbucket = int(math.ceil(float(self.slicesize) / maxnb))
+                lbucket = math.ceil(self.slicesize / maxnb)
         elif self.indsize == 2:
             # For light, we will never have to keep track of a
             # bucket outside of a block.
             maxnb = 2**16
             if self.blocksize > maxnb * lbucket:
-                lbucket = int(math.ceil(float(self.blocksize) / maxnb))
+                lbucket = math.ceil(self.blocksize / maxnb)
         else:
             # For medium and full indexes there should not be a need to
             # increase lbucket
@@ -970,7 +970,7 @@ class Index(NotLoggedMixin, Group, indexesextension.Index):
             message = f"swap_{what}"
         (nover, mult, tover) = self.compute_overlaps(
             self.tmp, message, self.verbose)
-        rmult = len(mult.nonzero()[0]) / float(len(mult))
+        rmult = len(mult.nonzero()[0]) / len(mult)
         if self.verbose:
             print(f"time: {time() - t1:.4f}. clock: {perf_counter() - c1:.4f}")
         # Check that entropy is actually decreasing
@@ -2006,8 +2006,8 @@ class Index(NotLoggedMixin, Group, indexesextension.Index):
         nslices = self.nslices
         lbucket = self.lbucket
         indsize = self.indsize
-        bucketsinblock = float(self.blocksize) / lbucket
-        nchunks = int(math.ceil(float(self.nelements) / lbucket))
+        bucketsinblock = self.blocksize / lbucket
+        nchunks = math.ceil(self.nelements / lbucket)
         chunkmap = numpy.zeros(shape=nchunks, dtype="bool")
         reduction = self.reduction
         starts = (self.starts - 1) * reduction + 1
@@ -2041,9 +2041,9 @@ class Index(NotLoggedMixin, Group, indexesextension.Index):
         if lbucket != nrowsinchunk:
             # Map the 'coarse grain' chunkmap into the 'true' chunkmap
             nelements = self.nelements
-            tnchunks = int(math.ceil(float(nelements) / nrowsinchunk))
+            tnchunks = math.ceil(nelements / nrowsinchunk)
             tchunkmap = numpy.zeros(shape=tnchunks, dtype="bool")
-            ratio = float(lbucket) / nrowsinchunk
+            ratio = lbucket / nrowsinchunk
             idx = chunkmap.nonzero()[0]
             starts = (idx * ratio).astype('int_')
             stops = numpy.ceil((idx + 1) * ratio).astype('int_')
